+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Added binding `ydb.WithWideTypes()` which interprets `time.Time` and `time.Duration` as `Timestamp64` and `Interval64` YDB types
+* Added binding `ydb.WithWideTimeTypes()` which interprets `time.Time` and `time.Duration` as `Timestamp64` and `Interval64` YDB types
 
 ## v3.103.0
 * Supported wide `Interval64` type